[core] Allow SpeedMonitor to handle very fast speeds
Over a decade ago speeds in excess of 2GB/sec would've been
fairly unthinkable, but nowadays that's not too hard to
imagine.

As the SpeedMonitor is an averager, and since we're using a
fairly naive implementation which sums everything up and then
divides by the averaging period, we can easily overflow 32bits
if the rate gets above 150MB/sec as the default averaging period
is 12.

Rather than trying to squeeze into 32bits, just make all the
internals 64bit as it may become a requirement in the not too
distant future.

+++ src/MonoTorrent/MonoTorrent.Client/Managers/DiskManager.cs
@@ -121,7 +121,7 @@ namespace MonoTorrent.Client
         /// <summary>
         /// The amount of data, in bytes, being read per second.
         /// </summary>
-        public int ReadRate => ReadMonitor.Rate;
+        public long ReadRate => ReadMonitor.Rate;
 
         /// <summary>
         /// The settings object passed to the ClientEngine, used to get the current read/write limits.
@@ -146,7 +146,7 @@ namespace MonoTorrent.Client
         /// <summary>
         /// The amount of data, in bytes, being written per second.
         /// </summary>
-        public int WriteRate => WriteMonitor.Rate;
+        public long WriteRate => WriteMonitor.Rate;
 
         /// <summary>
         /// The total number of bytes which have been read.
